accounts app ready for deletion, docker compose now reloads, users app is in use
