--- accounts/urls.py
@@ -1,8 +1,0 @@
-from django.urls import path
-
-from accounts import views
-
-
-urlpatterns = [
-    path("signup/", views.SignupPageView.as_view(), name="signup"),
-]
