+++ bookstore_project/urls.py
@@ -1,18 +1,3 @@
-"""bookstore_project URL Configuration
-
-The `urlpatterns` list routes URLs to views. For more information please see:
-    https://docs.djangoproject.com/en/2.2/topics/http/urls/
-Examples:
-Function views
-    1. Add an import:  from my_app import views
-    2. Add a URL to urlpatterns:  path('', views.home, name='home')
-Class-based views
-    1. Add an import:  from other_app.views import Home
-    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
-Including another URLconf
-    1. Import the include() function: from django.urls import include, path
-    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
-"""
 from django.conf import settings
 from django.contrib import admin
 from django.urls import include, path
@@ -22,8 +7,11 @@ urlpatterns = [
     # Dango Admin
     path('admin/', admin.site.urls),
 
-    # User management
+    # User Authentication
     path('accounts/', include('allauth.urls')),
+
+    # User Profiles
+    path('users/', include('users.urls')),
 
     # Local apps
     path('', include('pages.urls')),
